requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
firebase-admin==6.4.0
google-generativeai>=0.8.0
pydantic==2.5.3
pydantic-settings==2.1.0
